Refactor folders in /resources
