+++ .idea/workspace.xml
@@ -5,36 +5,9 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="307973d8-6b1d-4468-a75c-f97b5f92f18f" name="Changes" comment="Refactored Helper Functions into /Resources">
-      <change afterPath="$PROJECT_DIR$/src/resources/createUser.js" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/resources/fetchCharacter.js" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/resources/updateScore.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/package.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/prisma/schema.prisma" beforeDir="false" afterPath="$PROJECT_DIR$/prisma/schema.prisma" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/brandroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/brandroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/captainroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/captainroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/carmenroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/carmenroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/clinic/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/clinic/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/cooperroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/cooperroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/doyleroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/doyleroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/dressingroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/dressingroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/hallway/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/hallway/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/maanroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/maanroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/princessroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/princessroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/romilyroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/romilyroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/app/seraphineroom/page.jsx" beforeDir="false" afterPath="$PROJECT_DIR$/src/app/seraphineroom/page.jsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/characters/fetchCharacter.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/api/prisma/hints/index.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/fetchChoices.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/postChoices.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/hints/[id].js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/score/[id].js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/score/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/api/prisma/score/index.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/score/updateScore.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/timer/updateTimer.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/resources/updateTimer.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/user/createUser.js" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/score/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/api/prisma/score/update/index.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/pages/api/prisma/user/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/pages/api/prisma/user/index.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/pages/api/rooms/fetchRoom.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/resources/fetchRoom.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/yarn.lock" beforeDir="false" afterPath="$PROJECT_DIR$/yarn.lock" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -68,6 +41,7 @@
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "WebServerToolWindowFactoryState": "false",
     "git-widget-placeholder": "feat/VG-26",
+    "last_opened_file_path": "/Users/eugenewee/WebstormProjects/virtual-games/src/pages/api/prisma/score/update",
     "node.js.detected.package.eslint": "true",
     "node.js.detected.package.tslint": "true",
     "node.js.selected.package.eslint": "(autodetect)",
@@ -78,7 +52,11 @@
   }
 }]]></component>
   <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/src/pages/api/prisma/score/update" />
+    </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/src/pages/api/prisma/score/create" />
       <recent name="$PROJECT_DIR$/src/resources" />
     </key>
   </component>
@@ -106,7 +84,21 @@
       <option name="project" value="LOCAL" />
       <updated>1686387012641</updated>
     </task>
-    <option name="localTasksCounter" value="3" />
+    <task id="LOCAL-00003" summary="Refactored Helper Functions into /Resources">
+      <created>1686402908751</created>
+      <option name="number" value="00003" />
+      <option name="presentableId" value="LOCAL-00003" />
+      <option name="project" value="LOCAL" />
+      <updated>1686402908751</updated>
+    </task>
+    <task id="LOCAL-00004" summary="Refactored Helper Functions into /Resources">
+      <created>1686402943570</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1686402943570</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
