+++ src/app/captainroom/page.jsx
@@ -2,7 +2,7 @@
 import styles from "./components/styles.module.css";
 import { Container, Text, Box } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
-import fetchRoom from "@/resources/fetchRoom";
+import fetchRoom from "@/resources/cloudinary/fetchRoom";
 import CldImage from "../components/CldImage";
 import Map from "../Map";
