+++ src/app/dressingroom/page.jsx
@@ -1,5 +1,5 @@
 "use client";
-import fetchRoom from "@/resources/fetchRoom";
+import fetchRoom from "@/resources/cloudinary/fetchRoom";
 import styles from "./components/styles.module.css";
 import { Container, Text, Box } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
