+++ src/app/princessroom/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 import styles from "./components/styles.module.css";
 import { Container, Text, Box } from "@chakra-ui/react";
-import fetchRoom from "@/resources/fetchRoom";
+import fetchRoom from "@/resources/cloudinary/fetchRoom";
 import { useEffect, useState } from "react";
 import CldImage from "../components/CldImage";
 import Map from "../Map";
