+++ src/resources/cloudinary/fetchRoom.js
