+++ src/resources/prisma/updateScore.js
